Fixed unsafe casting size_t to int

+++ src/lexical_analyzer/span.cpp
@@ -10,14 +10,14 @@ std::string Span::to_string() const {
     return "Span(" + std::to_string(line_num) + ", " + std::to_string(pos_begin) + ", " + std::to_string(pos_end) + ")";
 }
 
-int Span::get_line_num() const {
+size_t Span::get_line_num() const {
     return line_num;
 }
 
-int Span::get_pos_begin() const {
+size_t Span::get_pos_begin() const {
     return pos_begin;
 }
 
-int Span::get_pos_end() const {
+size_t Span::get_pos_end() const {
     return pos_end;
 }
